at_requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.1
Pillow==9.4.0
requests==2.28.2
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
boto3==1.26.64